Fixed Conflicts
Fixed a few remaining conflicts in the addEvent() function in the executive class. Still getting errors adding to LinkedList.

+++ Events.cpp
@@ -86,7 +86,7 @@ void Events::getInfo()
 				{
 					timeSlot[i][j].getAttendees();
 					names = timeSlot[i][j].getAttendees();
-					for (int l = 0; l < names.size(); l++)
+					for (size_t l = 0; l < names.size(); l++)
 					{
 						std::cout << names[l];
 					}
